# [monodroid] don't probe for .exe files