--- a/src/monodroid/jni/embedded-assemblies.hh
+++ b/src/monodroid/jni/embedded-assemblies.hh
@@ -29,12 +29,6 @@ namespace xamarin::android::internal {
 #if defined (DEBUG) || !defined (ANDROID)
 		static constexpr char override_typemap_entry_name[] = ".__override__";
 #endif
-		static constexpr const char *suffixes[] = {
-			"",
-			".dll",
-			".exe",
-		};
-		static constexpr size_t SUFFIXES_SIZE = sizeof(suffixes) / sizeof (suffixes[0]);
 
 	public:
 		/* filename is e.g. System.dll, System.dll.mdb, System.pdb */